Some code clean up and moved MainScheduler API into a separate Go file to make it easier to refer to it
